refactor(vscode): split previewPanel webview JS + extract pure helpers
The two giant files in vscode-extension/ were carrying logic that didn't
need to be inline. Two complementary moves:

`previewPanel.ts` (3122 -> 174 LoC). The class was a thin wrapper around
a 2950-line inline `<script>` template literal — the entire webview UI
lived as a JS string in TypeScript, which prettier never touched and
which couldn't be linted independently. Move the IIFE body verbatim to
`media/preview.js` and load it via `<script src=...>`. The single
TS-side interpolation (`earlyFeaturesEnabled`) threads through as an
IIFE parameter sourced from a `data-early-features` attribute on the
script tag. CSP `script-src 'nonce-…'` already allows nonced external
scripts, so no policy change. `.vscodeignore` already ships
`media/**`, so the .vsix is byte-equivalent in shape.

`extension.ts` (4575 -> 4506 LoC). Three pure helpers had no
`vscode.*` dependencies and no module-state coupling: `sameScope`,
`isSourceFile`/`isGeneratedOutputPath`/`isPreviewSourceFile`
(`pathPredicates.ts`), `formatRelativeShort` (`relativeTime.ts`,
with an injectable clock), and `isAntigravityHost` (`hostDetection.ts`,
with an injectable env). Each lands with a focused unit-test file
(34 new assertions across the three modules); fast suite goes from
320 to 354 tests, still 172ms.

Behavior unchanged. Webview JS body is moved verbatim — prettier is
deliberately not run over `media/**/*.js` in this PR so the diff
stays a verifiable cut-and-paste; a follow-up can opt the file in.
